Split geom package into multiple Python files
Geom functions and classes should be imported as 'from phi.geom import X'
Deprecate phi.geom.geometry
