+++ phi/geom/__init__.py
@@ -1,1 +1,5 @@
-from .geometry import Geometry, Sphere, box, AABox, union
+from ._geom_util import assert_same_rank
+from ._geom import Geometry
+from ._union import union  # Union is private
+from ._box import AABox, box
+from ._sphere import Sphere
